+++ commands/createtext.js
@@ -1,7 +1,7 @@
 const { addItem } = require("../savesystem");
 const getItemEmbed = require("../getItemEmbed");
-const { adminRoleIds } = require("../config.json");
-
+const { adminRoleIds, prefix } = require("../config.json");
+const openCollectors = {};
 /**
  *
  * @param {import("discord.js").Message} message
@@ -18,6 +18,9 @@ const main = async (message) => {
       `Your roles dont have the permissions to add a item`
     );
 
+  if (openCollectors[message.author.id])
+    openCollectors[message.author.id].stop("new");
+
   const item = {};
   item.type = "text";
 
@@ -31,13 +34,14 @@ const main = async (message) => {
   const collector = message.channel.createMessageCollector(filter, {
     time: 180000,
   });
-
+  openCollectors[message.author.id] = collector;
   collector.on("collect", async (msg) => {
+    if (msg.content.startsWith(prefix)) return;
     if (collecting == "name") {
       item.name = msg.content;
       item.id = msg.content.toLowerCase();
-      msg.delete();
-      msgToDelete.delete();
+      await msg.delete();
+      await msgToDelete.delete();
       collecting = "content";
       msgToDelete = await message.channel.send(
         "Type in the content of your text!"
@@ -46,13 +50,14 @@ const main = async (message) => {
     }
     if (collecting == "content") {
       item.content = msg.content;
-      msg.delete();
-      msgToDelete.delete();
+      await msg.delete();
+      await msgToDelete.delete();
       collector.stop("done");
     }
   });
 
   collector.on("end", (_, reason) => {
+    delete openCollectors[message.author.id];
     message.delete();
     if (reason == "done") {
       const saveId = saveItem(item);
